+++ week02/LAB/demo/src/main/java/com/example/demo/service/dto/request/MemberCreateDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.dto;
+package com.example.demo.service.dto.request;
 
 
 import com.example.demo.domain.enums.Part;
