+++ week02/LAB/demo/src/main/java/com/example/demo/service/dto/response/MemberFindDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.dto;
+package com.example.demo.service.dto.response;
 
 
 import com.example.demo.domain.Member;
